fix(v)
switch to weekly builds. currently broken due to changes in the v c-compiler. git clone builds, so i expect this will be fixed with the next weekly release.

it git clones the current compiler from another archive to build, so all past versions are strictly unbuildable (as is the most recent release).