Cargo fixes (#3546)
Fixes crash due to cURL needing a patch for Sonoma. We use the Apple vendored cURL to work around this but probs will need to patch cURL as well. We'll see.

Note this is notable in that we elected to use more of our own deps on macOS than eg. brew, but probs should approach some balance with that.