the shim doesn’t need explicit pkgx dep
This caused all git dev-envs to implicitly depend on pkgx which would then cause a potentially older pkgx than the system installed one to be in most working envs. lol

Longer term a better solution would be nice.

In general having git in most dev-envs sucks since its deps are so intrusive. Would be nice to statically build parts of it or something.